requirements.txt
discord.py
aiohttp
python-dotenv
PyNaCl
asyncpg
pytest